Password Generator.py
This Python script generates a random password based on user input for the length of the password. It uses the random.choices() method to select characters from the string module, which includes uppercase and lowercase letters, digits, and punctuation symbols. The script then joins these characters into a string, creating a random password of the specified length