Find a number of  vowels and consonants
This Python script counts the number of vowels and consonants in a user-input string. It prompts the user to enter a string, then iterates over each character in the string. If the character is a vowel (including both uppercase and lowercase vowels), it increments the count_vowels variable; otherwise, it increments the count_Cons variable for consonants. Finally, it prints the total count of vowels and consonants in the input string.